# services/car-service/src/main/resources/api.yml
openapi: 3.0.0

info:
  title: Protogen API
  version: 1.0.0Beta
  description: Protogen API, implemented with OpenAPI Codegen

paths:
  /cars:
    post:
      summary: Creates a car
      operationId: createCar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarDTO'
      responses:
        '201':
          description: Car added successfully

    get:
      summary: Retrieves all cars that a user has access to
      operationId: getAllCars
      responses:
        '200':
          description: Retrieve success response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CarDTO'

  /cars/id/{id}:
    get:
      summary: Retrieve a car by id number
      operationId: getCarById
      parameters:
        - name: id
          in: path
          description: id number of a car to retrieve
          required: true
          schema:
            type: number
      responses:
        '200':
          description: Retrieve success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarDTO'
        '404':
          description: Car not found

  /cars/vin/{vin}:
    get:
      summary: Retrieve a car by vin number
      operationId: getCarByVin
      parameters:
        - name: vin
          in: path
          description: vin number of a car to retrieve
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Retrieve success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarDTO'
        '404':
          description: Car not found

    delete:
      summary: Remove a car by vin number
      operationId: removeCar
      parameters:
        - name: vin
          in: path
          description: vin number of a car to remove
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Car removed successfully
        '404':
          description: Car not found

    put:
      summary: update a car by vin number
      operationId: updateCar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarDTO'
      parameters:
        - name: vin
          in: path
          description: vin number of a car to update
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Car updated successfully
        '404':
          description: Car not found

  /cars/make/{make}:
    get:
      summary: Retrieve a car by make
      operationId: getCarByMake
      parameters:
        - name: make
          in: path
          description: make of a car to retrieve
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Retrieve success response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CarDTO'
        '404':
          description: Car not found

  /cars/color/{color}:
    get:
      summary: Retrieve a car by color
      operationId: getCarByColor
      parameters:
        - name: color
          in: path
          description: color of a car to retrieve
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Retrieve success response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CarDTO'
        '404':
          description: Car not found

  /cars/year/{carYear}:
    get:
      summary: Retrieve a car by year
      operationId: getCarByCarYear
      parameters:
        - name: carYear
          in: path
          description: year of a car to retrieve
          required: true
          schema:
            type: number
      responses:
        '200':
          description: Retrieve success response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CarDTO'
        '404':
          description: Car not found

  /cars/search:
    get:
      summary: Retrieve cars by search query
      operationId: searchCar
      parameters:
        - name: id
          in: query
          description: id of a car to search
          required: false
          schema:
            type: number
        - name: vin
          in: query
          description: vin of a car to search
          required: false
          schema:
            type: string
        - name: make
          in: query
          description: make of a car to search
          required: false
          schema:
            type: string
        - name: model
          in: query
          description: model of a car to search
          required: false
          schema:
            type: string
        - name: carYear
          in: query
          description: year of a car to search
          required: false
          schema:
            type: number
        - name: color
          in: query
          description: color of a car to search
          required: false
          schema:
            type: string
        - name: engine
          in: query
          description: engine of a car to search
          required: false
          schema:
            type: string
        - name: transmission
          in: query
          description: transmission of a car to search
          required: false
          schema:
            type: string
        - name: fuelType
          in: query
          description: fuel type of a car to search
          required: false
          schema:
            type: string
        - name: mileage
          in: query
          description: mileage of a car to search
          required: false
          schema:
            type: number
        - name: price
          in: query
          description: price of a car to search
          required: false
          schema:
            type: number
        - name: ownerId
          in: query
          description: owner Id of a car to search
          required: false
          schema:
            type: number
        - name: features
          in: query
          description: features of a car to search
          required: false
          schema:
            type: array
            items:
              string
      responses:
        '200':
          description: Retrieve success response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CarDTO'
        '404':
          description: Car not found

components:
  schemas:
    CarDTO:
      type: object
      properties:
        carId:
          type: number
        vin:
          type: string
        make:
          type: string
        model:
          type: string
        carYear:
          type: number
        color:
          type: string
        engine:
          type: number
        transmission:
          type: string
        fuelType:
          type: string
        mileage:
          type: number
        price:
          type: number
        ownerId:
          type: number
        features:
          type: array